requirements.txt
python-telegram-bot==21.0.1
google-generativeai==0.7.2
Pillow==11.0.0
